Moved gql to external file
